[READ DESC PLS]Movement bools
Tweaked movement script so that booleans (properties) that allows to know the state of the player are working. Jumping bool and grounded not working (because jumping one needs grounded one).

Exhaust isn't working anymore. That is normal, I'm reworking the boolean system and clarifying everything and it'll be okay tomorrow normally.